Proyecto Final/Python/requirements.txt
python-telegram-bot==7.0.1
requests==2.18.1
flask==0.12.1
flask-ask
